Add props to farmcard
