+++ package.json
@@ -11,6 +11,7 @@
     "@testing-library/user-event": "^12.2.2",
     "airtable": "^0.10.1",
     "font-awesome": "^4.7.0",
+    "prop-types": "^15.7.2",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
